fix: 매일위키 에러페이지, not-found 페이지 수정
