+++ src/app/global-error.tsx
@@ -1,7 +1,8 @@
 'use client';
 
 import { myStyle } from '@/_styles/vars.css';
-import Error from './error';
+import UnexpectedErrorPage from '@/_components/Error/UnexpectedError';
+import MaeilMailNav from '@/_components/common/MaeilMailNav';
 
 // root layout에서 발생한 에러는 global-error에 의해 처리됨.
 // layout.tsx을 대체하므로 html, body 필요
@@ -9,7 +10,7 @@ export default function GlobalError({ error }: { error: Error }) {
   return (
     <html>
       <body className={myStyle}>
-        <Error error={error} />;
+        <UnexpectedErrorPage Nav={<MaeilMailNav />} error={error} />
       </body>
     </html>
   );
