+++ src/app/maeil-wiki/not-found.tsx
@@ -1,0 +1,5 @@
+import NotFoundPage from '@/_components/Error/NotFound';
+
+export default function NotFound() {
+  return <NotFoundPage />;
+}
